requirements.txt
yfinance==0.2.40
mplfinance==0.12.10b0
tensorflow==2.15.0
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
tqdm==4.66.4
